refactor(reference): integrate apidom-json-pointer package
Refs #1491
