+++ packages/apidom-reference/package.json
@@ -37,6 +37,7 @@
     "@babel/runtime-corejs3": "=7.18.0",
     "@swagger-api/apidom-ast": "^0.27.0",
     "@swagger-api/apidom-core": "^0.27.0",
+    "@swagger-api/apidom-json-pointer": "^0.27.0",
     "@swagger-api/apidom-ns-asyncapi-2": "^0.27.0",
     "@swagger-api/apidom-ns-openapi-3-1": "^0.27.0",
     "@swagger-api/apidom-parser-adapter-asyncapi-json-2": "^0.27.0",
