+++ packages/apidom-reference/src/dereference/strategies/openapi-3-1/selectors/uri/index.ts
@@ -1,13 +1,10 @@
 import { isUndefined } from 'ramda-adjunct';
 import { Element, find, isStringElement } from '@swagger-api/apidom-core';
 import { isSchemaElement } from '@swagger-api/apidom-ns-openapi-3-1';
+import { uriToPointer, evaluate as jsonPointerEvaluate } from '@swagger-api/apidom-json-pointer';
 
 import * as url from '../../../../../util/url';
 import { EvaluationJsonSchemaUriError } from './errors';
-import {
-  uriToPointer,
-  evaluate as jsonPointerEvaluate,
-} from '../../../../../selectors/json-pointer';
 import { isAnchor, uriToAnchor, evaluate as $anchorEvaluate } from '../$anchor';
 
 // evaluates JSON Schema $ref containing unknown URI against ApiDOM fragment
