--- packages/apidom-reference/src/selectors/json-pointer/errors/EvaluationJsonPointerError.ts
@@ -1,1 +1,0 @@
-export default class EvaluationJsonPointerError extends Error {}
