--- packages/apidom-reference/src/selectors/json-pointer/errors/index.ts
@@ -1,2 +1,0 @@
-export { default as EvaluationJsonPointerError } from './EvaluationJsonPointerError';
-export { default as InvalidJsonPointerError } from './InvalidJsonPointerError';
